+++ java/src/main/java/com/leothenardo/homebroker/orders/application/ExecuteTransactionServiceInputDTO.java
@@ -7,7 +7,7 @@ public record ExecuteTransactionServiceInputDTO(
 				String sellerId,
 				String brokerTransactionId,
 				int negotiatedShares,
-				Float price,
+				Double price,
 				String assetId
 ) {
 }
